add hint for required canonical property names
When connecting to the notify signal for a property, we have to use the canonical property name. This is in particular relevant for typical python style underscore_case named properties, e.g., `foo_bar_baz`, where the canonical name replaces the underscore with a dash, i.e., the signal is `"notify::foo-bar-baz"`. Cost me some time and a false Bug report to PyGObject to figure that one out ... so maybe add it here? Think that might save some people some work in the future :).