Add a widget gallery page
This came up in https://gitlab.gnome.org/GNOME/pygobject/issues/208

Add a list of all non-abstract Gtk.Widget subclasses and show an
example image for each of them.
The images are taken from the gtk+ documentation.

The goal is that users can quickly get an overview of the available widgets
and to make it possible to find a widget type without knowing the name.